refactor CPS page composition
