+++ src/app/containers/ATIAnalytics/index.test.jsx
@@ -4,11 +4,10 @@ import { render } from '@testing-library/react';
 import { isNull, suppressPropWarnings } from '@bbc/psammead-test-helpers';
 import { articleDataNews } from '#pages/Article/fixtureData';
 import mapAssetData from '#pages/MediaAssetPage/fixtureData.json';
-import pglAssetData from '#pages/PhotoGallery/fixtureData.json';
-import styAssetData from '#pages/Story/fixtureData.json';
+import pglAssetData from '#pages/PhotoGalleryPage/fixtureData.json';
+import styAssetData from '#pages/StoryPage/fixtureData.json';
 import { RequestContextProvider } from '#contexts/RequestContext';
 import { ServiceContextProvider } from '#contexts/ServiceContext';
-
 import ATIAnalytics from '.';
 import * as amp from './amp';
 import * as canonical from './canonical';
