+++ src/app/pages/MediaAssetPage/index.jsx
@@ -1,6 +1,5 @@
 import React, { useContext } from 'react';
 import path from 'ramda/src/path';
-import pipe from 'ramda/src/pipe';
 import styled from 'styled-components';
 import {
   GEL_SPACING_DBL,
@@ -36,16 +35,9 @@ import {
 
 import { RequestContext } from '#contexts/RequestContext';
 
-// Page Handlers
-import withContexts from '#containers/PageHandlers/withContexts';
-import withPageWrapper from '#containers/PageHandlers/withPageWrapper';
-import withError from '#containers/PageHandlers/withError';
-import withLoading from '#containers/PageHandlers/withLoading';
-import withData from '#containers/PageHandlers/withData';
-
 const isLegacyMediaAssetPage = url => url.split('/').length > 7;
 
-const MediaAssetPageContainer = ({ pageData }) => {
+const MediaAssetPage = ({ pageData }) => {
   const requestContext = useContext(RequestContext);
   const title = path(['promo', 'headlines', 'headline'], pageData);
   const summary = path(['promo', 'summary'], pageData);
@@ -132,14 +124,6 @@ const MediaAssetPageContainer = ({ pageData }) => {
   );
 };
 
-MediaAssetPageContainer.propTypes = cpsAssetPagePropTypes;
+MediaAssetPage.propTypes = cpsAssetPagePropTypes;
 
-const EnhancedMediaAssetPageContainer = pipe(
-  withData,
-  withError,
-  withLoading,
-  withPageWrapper,
-  withContexts,
-)(MediaAssetPageContainer);
-
-export default EnhancedMediaAssetPageContainer;
+export default MediaAssetPage;
