+++ src/app/pages/PhotoGalleryPage/index.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import pipe from 'ramda/src/pipe';
 import styled from 'styled-components';
 import {
   GEL_SPACING_DBL,
@@ -30,14 +29,7 @@ import {
   getLastPublished,
 } from '#lib/utilities/parseAssetData';
 
-// Page Handlers
-import withContexts from '#containers/PageHandlers/withContexts';
-import withPageWrapper from '#containers/PageHandlers/withPageWrapper';
-import withError from '#containers/PageHandlers/withError';
-import withLoading from '#containers/PageHandlers/withLoading';
-import withData from '#containers/PageHandlers/withData';
-
-const PhotoGalleryPageContainer = ({ pageData }) => {
+const PhotoGalleryPage = ({ pageData }) => {
   const title = path(['promo', 'headlines', 'headline'], pageData);
   const summary = path(['promo', 'summary'], pageData);
   const metadata = path(['metadata'], pageData);
@@ -121,14 +113,6 @@ const PhotoGalleryPageContainer = ({ pageData }) => {
   );
 };
 
-PhotoGalleryPageContainer.propTypes = cpsAssetPagePropTypes;
+PhotoGalleryPage.propTypes = cpsAssetPagePropTypes;
 
-const EnhancedPhotoGalleryPageContainer = pipe(
-  withData,
-  withError,
-  withLoading,
-  withPageWrapper,
-  withContexts,
-)(PhotoGalleryPageContainer);
-
-export default EnhancedPhotoGalleryPageContainer;
+export default PhotoGalleryPage;
