+++ src/app/pages/PhotoGalleryPage/index.stories.jsx
@@ -2,24 +2,23 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { BrowserRouter } from 'react-router-dom';
 import WithTimeMachine from '#testHelpers/withTimeMachine';
-
-import StoryPage from '.';
+import { PhotoGalleryPage } from '..';
 import pageData from './fixtureData';
 
-const styStories = storiesOf('Pages|Story Page', module);
+const pglStories = storiesOf('Pages|Photo Gallery Page', module);
 
-styStories.addDecorator(story => <WithTimeMachine>{story()}</WithTimeMachine>);
+pglStories.addDecorator(story => <WithTimeMachine>{story()}</WithTimeMachine>);
 
-styStories.add('Mundo', () => {
+pglStories.add('Azeri', () => {
   return (
     <BrowserRouter>
-      <StoryPage
-        pageType="STY"
+      <PhotoGalleryPage
+        pageType="PGL"
         isAmp={false}
         pathname="/path"
         status={200}
         pageData={pageData}
-        service="mundo"
+        service="azeri"
       />
     </BrowserRouter>
   );
