+++ src/app/pages/StoryPage/categoryMap/index.js
