+++ src/app/pages/StoryPage/categoryMap/index.test.js
