+++ src/app/pages/index.js
@@ -1,12 +1,15 @@
 import pipe from 'ramda/src/pipe';
+import _MediaAssetPage from './MediaAssetPage';
+import _PhotoGalleryPage from './PhotoGalleryPage';
 import _RadioPage from './RadioPage';
+import _StoryPage from './StoryPage';
 import withContexts from '../containers/PageHandlers/withContexts';
 import withPageWrapper from '../containers/PageHandlers/withPageWrapper';
 import withError from '../containers/PageHandlers/withError';
 import withLoading from '../containers/PageHandlers/withLoading';
 import withData from '../containers/PageHandlers/withData';
 
-const enhancePage = pipe(
+const applyBasicPageHandlers = pipe(
   withData,
   withError,
   withLoading,
@@ -14,5 +17,7 @@ const enhancePage = pipe(
   withContexts,
 );
 
-// eslint-disable-next-line import/prefer-default-export
-export const RadioPage = enhancePage(_RadioPage);
+export const MediaAssetPage = applyBasicPageHandlers(_MediaAssetPage);
+export const PhotoGalleryPage = applyBasicPageHandlers(_PhotoGalleryPage);
+export const RadioPage = applyBasicPageHandlers(_RadioPage);
+export const StoryPage = applyBasicPageHandlers(_StoryPage);
